+++ packages/sdks-tests/src/e2e-tests/dynamic-loading.spec.ts
@@ -1,9 +1,13 @@
 import { expect } from '@playwright/test';
 import { test } from '../helpers/index.js';
 
+const checkSupportsDynamicLoading = (packageName: string) => {
+  test.skip(packageName !== 'sveltekit');
+};
+
 test.describe('Dynamic loading of custom components', () => {
   test('LazyComponent only loads when requested', async ({ page, packageName }) => {
-    test.skip(packageName !== 'sveltekit');
+    checkSupportsDynamicLoading(packageName);
     await page.goto('/dynamic-loading');
 
     let lazyComponentRequestedBeforeClick = false;
@@ -30,7 +34,7 @@ test.describe('Dynamic loading of custom components', () => {
     page,
     packageName,
   }) => {
-    test.skip(packageName !== 'sveltekit');
+    checkSupportsDynamicLoading(packageName);
     await page.goto('/dynamic-loading');
 
     let notLazyComponentRequested = false;
@@ -53,5 +57,46 @@ test.describe('Dynamic loading of custom components', () => {
     await page.waitForLoadState('networkidle');
 
     expect(notLazyComponentRequested).toBe(true);
+  });
+});
+
+test.describe('Eager dynamic loading of custom components', () => {
+  test('EagerLazyComponent loads immediately even if not requested', async ({
+    page,
+    packageName,
+  }) => {
+    checkSupportsDynamicLoading(packageName);
+    await page.goto('/eager-dynamic-loading');
+
+    let eagerLazyComponentRequested = false;
+    let eagerLazyComponentRendered = false;
+
+    page.on('request', request => {
+      if (request.url().includes('LazyComponent')) {
+        eagerLazyComponentRequested = true;
+      }
+    });
+
+    page.on('requestfinished', request => {
+      void (async () => {
+        if (request.resourceType() === 'document' || request.resourceType() === 'script') {
+          const response = await request.response();
+          const responseBody = await response?.text();
+
+          if (responseBody?.includes('lazy component loaded')) {
+            // component is SSR'd
+            eagerLazyComponentRendered = true;
+          }
+        }
+      })();
+    });
+
+    await page.waitForLoadState('networkidle');
+
+    expect(eagerLazyComponentRequested).toBe(true);
+    expect(eagerLazyComponentRendered).toBe(true);
+
+    const eagerLazyComponent = page.locator('text=lazy component loaded');
+    await expect(eagerLazyComponent).toBeVisible();
   });
 });