+++ packages/sdks-tests/src/specs/index.ts
@@ -62,6 +62,8 @@ import type { Sdk } from '../helpers/sdk.js';
 import { SYMBOL_WITH_REPEAT_INPUT_BINDING } from './symbol-with-repeat-input-binding.js';
 import { CUSTOM_COMPONENT_CHILDREN_SLOT_PLACEMENT } from './children-slot-placement.js';
 import { DYNAMIC_LOADING_CUSTOM_COMPONENTS } from './dynamic-loading.js';
+import { SSR_BINDING_CONTENT } from './ssr-binding.js';
+import { EAGER_DYNAMIC_LOADING_CUSTOM_COMPONENTS } from './eager-dynamic-loading.js';
 
 function isBrowser(): boolean {
   return typeof window !== 'undefined' && typeof document !== 'undefined';
@@ -138,9 +140,13 @@ export const PAGES = {
   '/custom-components-models-show': CUSTOM_COMPONENTS_MODELS_RESTRICTION,
   '/custom-components-models-not-show': CUSTOM_COMPONENTS_MODELS_RESTRICTION,
   '/editing-box-columns-inner-layout': EDITING_BOX_TO_COLUMN_INNER_LAYOUT,
+  '/get-content': HTTP_REQUESTS,
+  '/get-query': HTTP_REQUESTS,
   '/symbol-with-repeat-input-binding': SYMBOL_WITH_REPEAT_INPUT_BINDING,
   '/children-slot-placement': CUSTOM_COMPONENT_CHILDREN_SLOT_PLACEMENT,
   '/dynamic-loading': DYNAMIC_LOADING_CUSTOM_COMPONENTS,
+  '/eager-dynamic-loading': EAGER_DYNAMIC_LOADING_CUSTOM_COMPONENTS,
+  '/ssr-binding': SSR_BINDING_CONTENT,
 } as const;
 
 const apiVersionPathToProp = {
@@ -150,7 +156,12 @@ const apiVersionPathToProp = {
 
 export type Path = keyof typeof PAGES;
 
-const GEN1_ONLY_PATHNAMES: Path[] = ['/api-version-v1', '/personalization-container'];
+const GEN1_ONLY_PATHNAMES: Path[] = [
+  '/api-version-v1',
+  '/personalization-container',
+  '/get-query',
+  '/get-content',
+];
 const GEN2_ONLY_PATHNAMES: Path[] = [];
 
 export const getAllPathnames = (target: 'gen1' | 'gen2'): string[] => {
@@ -242,6 +253,16 @@ export const getProps = async (args: {
     case '/react-native-strict-style-mode':
       extraProps = {
         strictStyleMode: true,
+      };
+      break;
+    case '/get-content':
+      extraProps = {
+        options: { apiEndpoint: 'content' },
+      };
+      break;
+    case '/get-query':
+      extraProps = {
+        options: { apiEndpoint: 'query', format: 'html', model: 'abcd', key: 'abcd' },
       };
       break;
     case '/symbol-with-repeat-input-binding':