+++ packages/sdks/snippets/angular/src/app/announcement-bar/announcement-bar.component.ts
@@ -4,11 +4,18 @@
  * src/app/announcement-bar/announcement-bar.component.ts
  */
 
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { fetchOneEntry, type BuilderContent } from '@builder.io/sdk-angular';
+import {
+  Content,
+  fetchOneEntry,
+  type BuilderContent,
+} from '@builder.io/sdk-angular';
 
 @Component({
   selector: 'app-announcement-bar',
+  standalone: true,
+  imports: [CommonModule, Content],
   template: `
     <ng-container *ngIf="content">
       <builder-content