+++ packages/sdks/snippets/angular/src/app/app.module.ts
@@ -6,29 +6,34 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
-import { Content } from '@builder.io/sdk-angular';
 import { AnnouncementBarComponent } from './announcement-bar/announcement-bar.component';
 import { AppComponent } from './app.component';
 import { BlogArticleComponent } from './blog-article/blog-article.component';
 import { CatchAllComponent } from './catch-all/catch-all.component';
+import { NavBarComponent } from './nav-bar/nav-bar.component';
+import { ProductDetailsComponent } from './product-details/product-details.component';
 import { ProductEditorialComponent } from './product-editorial/product-editorial.component';
 
 @NgModule({
-  declarations: [AppComponent, AnnouncementBarComponent, CatchAllComponent],
-  // add Content to imports
+  declarations: [AppComponent],
   imports: [
     BrowserModule,
-    Content,
+    AnnouncementBarComponent,
     BlogArticleComponent,
     ProductEditorialComponent,
+    ProductDetailsComponent,
+    AnnouncementBarComponent,
+    NavBarComponent,
+    CatchAllComponent,
     RouterModule.forRoot([
       { path: 'announcements/:id', component: AnnouncementBarComponent },
       { path: 'blogs/new-product-line', component: BlogArticleComponent },
       { path: 'products/:id', component: ProductEditorialComponent },
+      { path: 'product/category/jacket', component: ProductDetailsComponent },
+      { path: 'landing-page', component: NavBarComponent },
       { path: '**', component: CatchAllComponent },
     ]),
   ],
-  providers: [],
   bootstrap: [AppComponent],
 })
 export class AppModule {}