+++ packages/sdks/snippets/angular/src/app/product-editorial/header/header.component.ts
@@ -1,10 +1,8 @@
-import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-header',
   standalone: true,
-  imports: [CommonModule],
   template: `
     <div
       style="display: flex;