+++ packages/sdks/src/functions/evaluate/edge-runtime/edge-runtime.ts
@@ -1,8 +1,7 @@
-import { logger } from '../../../helpers/logger';
-import { flattenState } from '../browser-runtime/browser';
-import type { ExecutorArgs } from '../helpers';
-import { getFunctionArguments } from '../helpers';
-import Interpreter from './acorn-interpreter';
+import { logger } from '../../../helpers/logger.js';
+import type { ExecutorArgs } from '../helpers.js';
+import { flattenState, getFunctionArguments } from '../helpers.js';
+import Interpreter from './acorn-interpreter.js';
 
 /**
  * https://stackoverflow.com/a/46503625
@@ -41,8 +40,8 @@ function patchInterpreter() {
     obj._connected[name] = this.pseudoToNative(value);
   }
 
-  const getKeys = [];
-  const setKeys = [];
+  const getKeys: string[] = [];
+  const setKeys: string[] = [];
   for (const key of Object.keys(Interpreter.prototype)) {
     if (Interpreter.prototype[key] === originalGetProperty) {
       getKeys.push(key);
@@ -67,17 +66,19 @@ function patchInterpreter() {
 }
 patchInterpreter();
 
-const stripAwaiter = (code: string) => {
-  // remove anything before 'function main()'
-  return code.replace(
-    /^.*?function main\(\)/,
-    `var __awaiter = function (e, t, n, r) {
-    return r()
-  },
-  __generator = function (e, t) {
-    return t()
-  };
-  function main()`
+const processCode = (code: string) => {
+  return (
+    code
+      // strip async/await polyfill: remove anything before 'function main()'
+      .replace(
+        /^.*?function main\(\)/,
+        `
+var __awaiter = function (e, t, n, r) {return r()},
+__generator = function (e, t) { return t() };
+function main()`
+      )
+      // replace ?. with .
+      .replace(/\?\./g, '.')
   );
 };
 
@@ -116,7 +117,7 @@ export const runInEdge = ({
       return `var ${key} = ${jsonValName} === undefined ? undefined : JSON.parse(${jsonValName});`;
     })
     .join('\n');
-  const cleanedCode = stripAwaiter(code);
+  const cleanedCode = processCode(code);
   if (cleanedCode === '') {
     logger.warn('Skipping evaluation of empty code block.');
     return;